backend/requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.18.0
fastapi==0.68.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
motor==3.3.2
passlib==1.7.4
pip==23.3.1
pyasn1==0.5.1
pydantic==1.10.14
PyJWT==1.7.1
pymongo==4.6.2
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
setuptools==68.2.2
six==1.16.0
starlette==0.14.2
typing_extensions==4.10.0
uvicorn==0.15.0
sqlalchemy==1.4
psycopg2-binary==2.9.5
requests==2.31.0
